Refactor

+++ tests/conftest.py
@@ -1,11 +1,19 @@
 import pytest
-import pymongo
+
+try:
+    import pymongo
+    PYMONGO_INSTALLED = True
+except:
+    PYMONGO_INSTALLED = False
 
 
 @pytest.fixture(autouse=True)
 def setup_mongodb_database():
-    client = pymongo.MongoClient()
-    db = client["metadata_test"]
-    yield db
-    client.drop_database(db)
-    client.close()
+    if not PYMONGO_INSTALLED:
+        yield False
+    else:
+        client = pymongo.MongoClient()
+        db = client["metadata_test"]
+        yield db
+        client.drop_database(db)
+        client.close()
